Move futures account toggle to separate component
